server/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Coders for Causes"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
psycopg = {extras = ["binary", "pool"], version = "^3.1.19"}
freezegun = "^1.5.1"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"
setuptools = "^70.2.0"
pillow = "^10.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
flake8 = "6.1.0"
flake8-django = "^1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
